# docs/api/swagger.yaml
basePath: /v1
definitions:
...
  response.Envelope:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      success:
        type: boolean
    type: object
  response.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  user.UpdateProfileInput:
    properties:
      countryCode:
        type: string
      displayName:
        type: string
      email:
        type: string
      phone:
        type: string
      preferredLanguage:
        type: string
    type: object
host: moistello.com
info:
  contact:
    email: support@moistello.com
    name: Moistello Support
    url: https://moistello.com/support
  description: Decentralized savings circles on Stellar. REST API for circles, contributions,
    payouts, reputation, and governance.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://moistello.com/terms
  title: Moistello API
  version: 1.0.0
paths:
...
  /notifications:
    get:
      description: Returns paginated notifications for the authenticated user. Use
        ?unread=true to filter unread only.
      parameters:
      - description: Filter unread only
        in: query
        name: unread
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    notifications:
                      type: array
                  type: object
                meta:
                  $ref: '#/definitions/response.PaginationMeta'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - Notifications
  /notifications/{id}/read:
    patch:
      description: Marks a single notification as read.
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    success:
                      type: boolean
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - Notifications
  /notifications/preferences:
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's notification channel preferences
        and mute status.
      parameters:
      - description: Preferences
        in: body
        name: body
        required: true
        schema:
          properties:
            channels:
              type: array
            muted:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    preferences:
                      type: object
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Update notification preferences
      tags:
      - Notifications
  /notifications/read-all:
    patch:
      description: Marks every unread notification for the authenticated user as read.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    success:
                      type: boolean
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - Notifications
...
  /users/{id}:
    get:
      description: Returns a user's public profile by wallet address or user ID.
      parameters:
      - description: User ID or wallet address
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    user:
                      type: object
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Get user by ID
      tags:
      - Users
  /users/me:
    get:
      description: Returns the authenticated user's profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    user:
                      type: object
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's profile fields (displayName, email,
        countryCode, preferredLanguage).
      parameters:
      - description: Profile updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.UpdateProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    user:
                      type: object
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - Users
  /users/me/circles:
    get:
      description: Returns all circles the authenticated user belongs to.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    circles:
                      type: array
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get my circles
      tags:
      - Users
  /users/me/kyc:
    post:
      description: Starts the KYC verification process for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    kycStatus:
                      type: string
                    user:
                      type: object
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Initiate KYC verification
      tags:
      - Users
  /users/me/kyc/status:
    get:
      description: Returns the current KYC verification status of the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    kycStatus:
                      type: string
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get KYC status
      tags:
      - Users
  /users/me/reputation:
    get:
      description: Returns the authenticated user's MoiScore reputation.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    reputation:
                      type: number
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get my reputation score
      tags:
      - Users
  /webhooks:
    get:
      description: Lists all registered webhooks for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    webhooks:
                      type: array
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: List webhooks
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Registers a new webhook endpoint to receive event notifications.
      parameters:
      - description: Webhook config
        in: body
        name: body
        required: true
        schema:
          properties:
            events:
              type: array
            url:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    webhook:
                      type: object
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Register a webhook
      tags:
      - Webhooks
  /webhooks/{id}:
    delete:
      description: Deletes a registered webhook by ID.
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    success:
                      type: boolean
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Delete a webhook
      tags:
      - Webhooks
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from /auth/verify or /auth/register. Format:
      "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"